pyproject.toml
[project]
name = "converter"
version = "0.1.0"
description = ""
authors = [
    {name = "darkentrery",email = "badajkovmihail@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram[polling] (>=3.17.0,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "pydantic (>=2.10.6,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
